- rename max & min priority queues files
