+++ src/adts/max-priority-queue/max-priority-queue.client.js
@@ -1,43 +1,43 @@
-const MinPQ = require('./priority-queue-min')
+const MaxPQ = require('./max-priority-queue')
 const { StdIn, StdOut } = require('../../libs')
 const { Stack, Transaction } = require('../index')
 
 /**
- * TopM
- * @classdesc Min Priority Queue Test client.
+ * LowestM
+ * @classdesc Max Priority Queue Test client.
  * @see p. 311
  */
-class TopM {
+class LowestM {
   /**
-   * Prints the top `m` transactions from the StdIn.
+   * Prints the lowest `m` transactions from the StdIn.
    * @param {[]} args [m]
-   * @example <caption>Print the top 5 transactions</caption>
+   * @example <caption>Print the lowest 5 transactions</caption>
    * ```sh
-   * $ node priority-queue-min.client.js 5 < ~/algs4-data/tinyBatch.txt
-   * Thompson 2/27/2000 4747.08
-   * vonNeumann 2/12/1994 4732.35
-   * vonNeumann 1/11/1999 4409.74
-   * Hoare 8/18/1992 4381.21
-   * vonNeumann 3/26/2002 4121.85
+   * $ node max-priority-queue.client.js 5 < ~/algs4-data/tinyBatch.txt
+   * Turing 1/11/2002 66.1
+   * Turing 6/17/1990 644.08
+   * Dijkstra 9/10/2000 708.95
+   * Dijkstra 11/18/1995 837.42
+   * Hoare 8/12/2003 1025.7
    * ```
    */
   static main (args) {
     const m = parseInt(args[0], 10)
-    const pq = new MinPQ(m + 1)
+    const pq = new MaxPQ(m + 1)
 
     StdIn.read()
       .on('line', line => {
         pq.insert(new Transaction(line.trim()))
 
         if (pq.size() > m) {
-          pq.delMin()
+          pq.delMax()
         }
       })
       .on('close', () => {
         const stack = new Stack()
 
         while (!pq.isEmpty()) {
-          stack.push(pq.delMin())
+          stack.push(pq.delMax())
         }
 
         for (const transaction of stack) {
@@ -49,4 +49,4 @@ class TopM {
 
 // Execution
 // ==============================
-TopM.main(process.argv.slice(2))
+LowestM.main(process.argv.slice(2))
