+++ src/adts/min-priority-queue/min-priority-queue.client.js
@@ -1,43 +1,43 @@
-const MaxPQ = require('./priority-queue-max')
+const MinPQ = require('./min-priority-queue')
 const { StdIn, StdOut } = require('../../libs')
 const { Stack, Transaction } = require('../index')
 
 /**
- * LowestM
- * @classdesc Max Priority Queue Test client.
+ * TopM
+ * @classdesc Min Priority Queue Test client.
  * @see p. 311
  */
-class LowestM {
+class TopM {
   /**
-   * Prints the lowest `m` transactions from the StdIn.
+   * Prints the top `m` transactions from the StdIn.
    * @param {[]} args [m]
-   * @example <caption>Print the lowest 5 transactions</caption>
+   * @example <caption>Print the top 5 transactions</caption>
    * ```sh
-   * $ node priority-queue-max.client.js 5 < ~/algs4-data/tinyBatch.txt
-   * Turing 1/11/2002 66.1
-   * Turing 6/17/1990 644.08
-   * Dijkstra 9/10/2000 708.95
-   * Dijkstra 11/18/1995 837.42
-   * Hoare 8/12/2003 1025.7
+   * $ node min-priority-queue.client.js 5 < ~/algs4-data/tinyBatch.txt
+   * Thompson 2/27/2000 4747.08
+   * vonNeumann 2/12/1994 4732.35
+   * vonNeumann 1/11/1999 4409.74
+   * Hoare 8/18/1992 4381.21
+   * vonNeumann 3/26/2002 4121.85
    * ```
    */
   static main (args) {
     const m = parseInt(args[0], 10)
-    const pq = new MaxPQ(m + 1)
+    const pq = new MinPQ(m + 1)
 
     StdIn.read()
       .on('line', line => {
         pq.insert(new Transaction(line.trim()))
 
         if (pq.size() > m) {
-          pq.delMax()
+          pq.delMin()
         }
       })
       .on('close', () => {
         const stack = new Stack()
 
         while (!pq.isEmpty()) {
-          stack.push(pq.delMax())
+          stack.push(pq.delMin())
         }
 
         for (const transaction of stack) {
@@ -49,4 +49,4 @@ class LowestM {
 
 // Execution
 // ==============================
-LowestM.main(process.argv.slice(2))
+TopM.main(process.argv.slice(2))
